+++ tree-sitter-stack-graphs/src/cli/parse.rs
@@ -40,14 +40,14 @@ impl ParseArgs {
                 None => return Err(anyhow!("No stack graph language found")),
             };
         let source = file_reader.get(&self.source_path)?;
-        let tree = parse(lang, &self.source_path, source)?;
+        let tree = parse(&lang, &self.source_path, source)?;
         print_tree(tree);
         Ok(())
     }
 }
 
 pub(super) fn parse(
-    language: tree_sitter::Language,
+    language: &tree_sitter::Language,
     path: &Path,
     source: &str,
 ) -> anyhow::Result<tree_sitter::Tree> {