+++ tree-sitter-stack-graphs/src/loader.rs
@@ -68,13 +68,12 @@ impl LanguageConfiguration {
         builtins_config: Option<&str>,
         cancellation_flag: &dyn CancellationFlag,
     ) -> Result<Self, LoadError<'a>> {
-        let sgl = StackGraphLanguage::from_source(language, tsg_path.clone(), tsg_source).map_err(
-            |err| LoadError::SglParse {
-                inner: err,
-                tsg_path,
-                tsg: Cow::from(tsg_source),
-            },
-        )?;
+        let sgl = StackGraphLanguage::from_source(language.clone(), tsg_path.clone(), tsg_source)
+            .map_err(|err| LoadError::SglParse {
+            inner: err,
+            tsg_path,
+            tsg: Cow::from(tsg_source),
+        })?;
         let mut builtins = StackGraph::new();
         if let Some((builtins_path, builtins_source)) = builtins_source {
             let mut builtins_globals = Variables::new();
@@ -273,7 +272,7 @@ impl Loader {
         &mut self,
         path: &Path,
         content: &mut dyn ContentProvider,
-    ) -> Result<Option<tree_sitter::Language>, LoadError<'static>> {
+    ) -> Result<Option<&tree_sitter::Language>, LoadError<'static>> {
         match &mut self.0 {
             LoaderImpl::Paths(loader) => loader.load_tree_sitter_language_for_file(path, content),
             LoaderImpl::Provided(loader) => {
@@ -496,10 +495,10 @@ impl LanguageConfigurationsLoader {
         &mut self,
         path: &Path,
         content: &mut dyn ContentProvider,
-    ) -> Result<Option<tree_sitter::Language>, LoadError<'static>> {
+    ) -> Result<Option<&tree_sitter::Language>, LoadError<'static>> {
         for configuration in self.configurations.iter() {
             if configuration.matches_file(path, content)? {
-                return Ok(Some(configuration.language));
+                return Ok(Some(&configuration.language));
             }
         }
         Ok(None)
@@ -567,9 +566,9 @@ impl PathLoader {
         &mut self,
         path: &Path,
         content: &mut dyn ContentProvider,
-    ) -> Result<Option<tree_sitter::Language>, LoadError<'static>> {
+    ) -> Result<Option<&tree_sitter::Language>, LoadError<'static>> {
         if let Some(selected_language) = self.select_language_for_file(path, content)? {
-            return Ok(Some(selected_language.language));
+            return Ok(Some(&selected_language.language));
         }
         Ok(None)
     }
@@ -591,7 +590,7 @@ impl PathLoader {
             Some(index) => index,
             None => {
                 let tsg = self.load_tsg_from_paths(&language)?;
-                let sgl = StackGraphLanguage::new(language.language, tsg);
+                let sgl = StackGraphLanguage::new(language.language.clone(), tsg);
 
                 let mut builtins = StackGraph::new();
                 self.load_builtins_from_paths_into(
@@ -602,7 +601,7 @@ impl PathLoader {
                 )?;
 
                 let lc = LanguageConfiguration {
-                    language: language.language,
+                    language: language.language.clone(),
                     scope: language.scope,
                     content_regex: language.content_regex,
                     file_types: language.file_types,
@@ -696,7 +695,7 @@ impl PathLoader {
             }
             if tsg_path.exists() {
                 let tsg_source = std::fs::read_to_string(tsg_path)?;
-                return Loader::load_tsg(language.language, Cow::from(tsg_source));
+                return Loader::load_tsg(language.language.clone(), Cow::from(tsg_source));
             }
         }
         return Err(LoadError::NoTsgFound);
@@ -786,10 +785,11 @@ impl SupplementedTsLoader {
                 .map_err(LoadError::TreeSitter)?;
             let configurations = self
                 .0
-                .find_language_configurations_at_path(&path)
+                .find_language_configurations_at_path(&path, true)
                 .map_err(LoadError::TreeSitter)?;
             let languages = languages
                 .into_iter()
+                .map(|(l, _)| l)
                 .zip(configurations.into_iter())
                 .map(SupplementedLanguage::from)
                 .filter(|language| scope.map_or(true, |scope| language.matches_scope(scope)))