+++ tree-sitter-stack-graphs/tests/it/main.rs
@@ -28,7 +28,7 @@ pub(self) fn build_stack_graph(
 ) -> Result<(StackGraph, Handle<File>), BuildError> {
     let file_name = "test.py";
     let language =
-        StackGraphLanguage::from_str(tree_sitter_python::language(), tsg_source).unwrap();
+        StackGraphLanguage::from_str(tree_sitter_python::LANGUAGE.into(), tsg_source).unwrap();
     let mut graph = StackGraph::new();
     let file = graph.get_or_create_file(file_name);
     let mut globals = Variables::new();