+++ languages/tree-sitter-stack-graphs-java/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.5.0 -- 2024-12-12
+
+- The `tree-sitter-stack-graphs` dependency is updated to version 0.10.
+
+- The `tree-sitter-java` dependency is updated to version 0.23.4.
+
 ## v0.4.0 -- 2024-07-09
 
 ### Added